Converted article index to use a client component wrapper instead of using a mix which was throwing RSC payload errors
